+++ upload_no_build.py
@@ -1,0 +1,8 @@
+Import("env")
+env.AddCustomTarget(
+    "uploadnobuild", 
+    None, 
+    'pio run -e %s -t nobuild -t upload' %
+        env["PIOENV"], 
+    title="Upload without building"
+)